Moved view Vue files from components to views folder to improve organization and better align with standards
